Add support for strategy.matrix builds
The behaviour of conan and conan_package_tools
is conrolled by environment variables
and there is no way to undefine a variable from
a workflow file after it was set.

For example, you cannot set CONAN_DOCKER_IMAGE
and then unset it.

This does not allow you to use strategy.matrix
to run several jobs based on a single job description.

This commit adds several input arguments to overcome
this issue. If set these arguments override the
corresponding environment variables:
* compiler
* compiler-versions
* docker-image